dev/backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==3.7.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.104.1
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0
Pillow==10.1.0
